refactor(interpreter): move builtins to module
